Seperate code into a few files
