+++ core/classes/Endpoints/Endpoints.php
@@ -12,13 +12,11 @@
 
 class Endpoints {
 
+    use MatchesRoutes;
+    use ManagesTransformers;
+
     /** @var EndpointBase[] */
     private iterable $_endpoints = [];
-
-    /**
-     * @var array Mapping of key names to closures to transform a variable into an object (ie, a user ID to a User object)
-     */
-    private static array $_transformers = [];
 
     /**
      * Register an endpoint if it's route is not already taken.
@@ -40,15 +38,6 @@ class Endpoints {
      */
     public function getAll(): iterable {
         return $this->_endpoints;
-    }
-
-    /**
-     * Get all registered transformers
-     *
-     * @return array All transformers.
-     */
-    public static function getAllTransformers(): array {
-        return self::$_transformers;
     }
 
     /**
@@ -76,7 +65,7 @@ class Endpoints {
 
                     $reflection = new ReflectionMethod($endpoint, 'execute');
                     if ($reflection->getNumberOfParameters() !== (count($vars) + 1)) {
-                        throw new InvalidArgumentException("Endpoint's 'execute()' method must take " . (count($vars) + 1) . " arguments.");
+                        throw new InvalidArgumentException("Endpoint's 'execute()' method must take " . (count($vars) + 1) . " arguments. Endpoint: " . $endpoint->getRoute());
                     }
 
                     if (!$endpoint->isAuthorised($api)) {
@@ -100,99 +89,5 @@ class Endpoints {
         }
 
         $api->throwError(3, $api->getLanguage()->get('api', 'invalid_api_method'), 'If you are seeing this while in a browser, this does not mean your API is not functioning!', 404);
-    }
-
-    /**
-     * Determine if an Endpoint matches a route.
-     * If it does, return an array of variables to pass to the endpoint.
-     *
-     * @param EndpointBase $endpoint Endpoint to attempt to match.
-     * @param string $route Route to match.
-     * @return array|false Array of variables to pass to the endpoint, or false if the route does not match.
-     */
-    private function matchRoute(EndpointBase $endpoint, string $route) {
-        $endpoint_parts = explode('/', $endpoint->getRoute());
-        $endpoint_vars = [];
-
-        $route_parts = explode('/', $route);
-        $route_vars = [];
-
-        if (count($endpoint_parts) !== count($route_parts)) {
-            return false;
-        }
-
-        $i = 0;
-        // first, find any variables (e.g. {user}) in the endpoint's route
-        // we save them to an array with their index, so we can reference them later
-        foreach ($endpoint_parts as $part) {
-            if (strpos($part, '{') === 0 && substr($part, -1) === '}') {
-                $endpoint_vars[$i] = substr($part, 1, -1);
-            }
-            $i++;
-        }
-
-        $i = 0;
-        // now we go over the route and, if each piece is a variable (according to its index), add it to the returned variable array
-        // otherwise, if it's not supposed to be a variable, we check if it matches the endpoint's respective route fragment and exit if it doesn't
-        foreach ($route_parts as $part) {
-            if (array_key_exists($i, $endpoint_vars)) {
-                $route_vars[$endpoint_vars[$i]] = $part;
-            } else if ($endpoint_parts[$i] !== $part) {
-                return false;
-            }
-            $i++;
-        }
-
-        return $route_vars;
-    }
-
-    /**
-     * Register a transformer for API route binding.
-     *
-     * @param string $type The name of the transformer. This is used to identify the transformer when binding.
-     * @param string $module The name of the module that registered the transformer.
-     * @param Closure(Nameless2API, string): mixed $transformer Function which converts the value in the URL to the desired type.
-     */
-    public static function registerTransformer(string $type, string $module, Closure $transformer): void {
-        if (isset(self::$_transformers[$type])) {
-            throw new InvalidArgumentException("A transformer with for the type '$type' has already been registered by the '" . self::$_transformers[$type]['module'] . "' module.");
-        }
-
-        $reflection = new ReflectionFunction($transformer);
-        if ($reflection->getNumberOfParameters() !== 2) {
-            throw new InvalidArgumentException('Endpoint variable transformer must take 2 arguments (Nameless2API and the raw variable).');
-        }
-
-        // if they've provided a typehint for the first argument, make sure it's taking Nameless2API
-        $param = $reflection->getParameters()[0];
-        if ($param->getType() instanceof ReflectionNamedType && $param->getType()->getName() !== Nameless2API::class) {
-            throw new InvalidArgumentException('Endpoint variable transformer must take Nameless2API as the first argument.');
-        }
-
-        // check that the second argument is a string
-        $param = $reflection->getParameters()[1];
-        if ($param->getType() instanceof ReflectionNamedType && $param->getType()->getName() !== 'string') {
-            throw new InvalidArgumentException('Endpoint variable transformer must take a string as the second argument.');
-        }
-
-        self::$_transformers[$type] = [
-            'module' => $module,
-            'transformer' => $transformer,
-        ];
-    }
-
-    /**
-     * Convert a value through a transformer based on its type. If no transformer is found, the value is returned as-is.
-     *
-     * @param Nameless2API $api Instance of API to provide the transformer.
-     * @param string $type The type to use.
-     * @param string $value The value to convert.
-     */
-    private static function transform(Nameless2API $api, string $type, string $value) {
-        if (array_key_exists($type, self::$_transformers)) {
-            return self::$_transformers[$type]['transformer']($api, $value);
-        }
-
-        return $value;
     }
 }
